Simplified DistributionPolicy using concurrent dict.
