+++ src/NServiceBus.Core.Tests/Routing/SingleInstanceRoundRobinDistributionStrategyTests.cs
@@ -47,52 +47,19 @@
 
             var result = new List<UnicastRoutingTarget>();
             result.AddRange(strategy.SelectDestination(instances));
-            Enumerable.Range(1, 2).Select(_ => strategy.SelectDestination(instances)).ToList();
             result.AddRange(strategy.SelectDestination(instances));
-            Enumerable.Range(1, 2).Select(_ => strategy.SelectDestination(instances)).ToList();
+            result.AddRange(strategy.SelectDestination(instances));
             result.AddRange(strategy.SelectDestination(instances));
 
-            Assert.That(result, Has.All.SameAs(result.First()));
-        }
-
-        [Test]
-        public void ShouldScopeDistributionToEndpointName()
-        {
-            var strategy = new SingleInstanceRoundRobinDistributionStrategy();
-
-            var endpointA = "endpointA";
-            var endpointAInstances = new[]
-            {
-                UnicastRoutingTarget.ToEndpointInstance(new EndpointInstance(endpointA, "1")),
-                UnicastRoutingTarget.ToEndpointInstance(new EndpointInstance(endpointA, "2")),
-            };
-
-            var endpointB = "endpointB";
-            var endpointBInstances = new[]
-            {
-                UnicastRoutingTarget.ToEndpointInstance(new EndpointInstance(endpointB, "1")),
-                UnicastRoutingTarget.ToEndpointInstance(new EndpointInstance(endpointB, "2")),
-            };
-
-            var result = new List<UnicastRoutingTarget>();
-            result.AddRange(strategy.SelectDestination(endpointAInstances));
-            result.AddRange(strategy.SelectDestination(endpointBInstances));
-            result.AddRange(strategy.SelectDestination(endpointAInstances));
-            result.AddRange(strategy.SelectDestination(endpointBInstances));
-
-            Assert.That(result.Count, Is.EqualTo(4));
-            Assert.That(result, Has.Exactly(1).EqualTo(endpointAInstances[0]));
-            Assert.That(result, Has.Exactly(1).EqualTo(endpointAInstances[1]));
-            Assert.That(result, Has.Exactly(1).EqualTo(endpointBInstances[0]));
-            Assert.That(result, Has.Exactly(1).EqualTo(endpointBInstances[1]));
+            Assert.That(result.Last(), Is.EqualTo(result.First()));
         }
 
         [Test]
         public void WhenNewInstancesAdded_ShouldIncludeAllInstancesInDistribution()
         {
+            var endpointName = "endpointA";
             var strategy = new SingleInstanceRoundRobinDistributionStrategy();
 
-            var endpointName = "endpointA";
             var instances = new List<UnicastRoutingTarget>
             {
                 UnicastRoutingTarget.ToEndpointInstance(new EndpointInstance(endpointName, "1")),
